Holy Cow! I have decided that the time had come to do what I had said I was going to do this whole time, and I have spent the past 13-14 hours making a major push into getting a lot of pieces into place. A lot of the code is bare-bones currently and there are some issues that have come up that need urgent addressing, but I would say I have a nice framework in place here and it's all just details from this point on. One major issue is the way I have a debug/release version of the GLAD library being selected based off macros is NOT going to work unless I can figure out a way to put in some different headers or something.
Anyways, my I tested my callback-queue system with a joystick and it worked like a charm.

I have to be awake for a big day tomorrow starting in about 5 hours, but from what I was able to accomplish today has made it worth it.

It's only Uphill up the hill of progress from here